requirements.txt

asgiref==3.2.7
astroid==2.3.3
colorama==0.4.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-crispy-forms==1.9.0
django-import-export==2.0.2
et-xmlfile==1.0.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.1
pylint==2.4.4
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
typed-ast==1.4.1
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0